+++ pkg/helm/values.go
@@ -1,18 +1,28 @@
 package helm
 
 import (
+	"bytes"
+	"embed"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/url"
 	"os"
 	"strings"
+	textTemplate "text/template"
 
 	"github.com/imdario/mergo"
 	"gopkg.in/yaml.v3"
 )
 
-func GetChartValuesOverrides(paths []string) (map[string]interface{}, error) {
+//go:embed templates/*
+var templatesFS embed.FS
+
+type TemplateValues struct {
+	StorageClassName string
+}
+
+func GetChartValuesOverrides(paths []string, templateValues *TemplateValues) (map[string]interface{}, error) {
 	var err error
 
 	valuesOverride := make(map[string]interface{})
@@ -23,7 +33,7 @@ func GetChartValuesOverrides(paths []string) (map[string]interface{}, error) {
 		}
 
 		var data []byte
-		if data, err = readValuesOverride(path); err != nil {
+		if data, err = readValuesOverride(path, templateValues); err != nil {
 			return nil, err
 		}
 
@@ -40,8 +50,27 @@ func GetChartValuesOverrides(paths []string) (map[string]interface{}, error) {
 	return valuesOverride, nil
 }
 
-func readValuesOverride(path string) ([]byte, error) {
+func readTemplateOverride(path string, templateValues *TemplateValues) ([]byte, error) {
 	var err error
+	var template *textTemplate.Template
+	if template, err = textTemplate.ParseFS(templatesFS, path); err != nil {
+		return nil, err
+	}
+
+	buffer := new(bytes.Buffer)
+	if err = template.Execute(buffer, templateValues); err != nil {
+		return nil, err
+	}
+
+	return buffer.Bytes(), nil
+}
+
+func readValuesOverride(path string, templateValues *TemplateValues) ([]byte, error) {
+	var err error
+
+	if strings.HasPrefix(path, "templates") {
+		return readTemplateOverride(path, templateValues)
+	}
 
 	if strings.HasPrefix(path, "presets") {
 		return presetsFS.ReadFile(path)
@@ -59,7 +88,7 @@ func readValuesOverride(path string) ([]byte, error) {
 			return nil, fmt.Errorf("[%d] %s download failed", response.StatusCode, path)
 		}
 
-		return ioutil.ReadAll(response.Body)
+		return io.ReadAll(response.Body)
 	}
 
 	return os.ReadFile(path)